actually create all tables in tests

+++ src/tests/ggrc/__init__.py
@@ -19,7 +19,8 @@ if os.environ.get('TRAVIS', False):
 class TestCase(BaseTestCase):
   def setUp(self):
     for table in reversed(db.metadata.sorted_tables):
-      db.engine.execute(table.delete())
+      if not table.name == "test_model":
+        db.engine.execute(table.delete())
     db.session.commit()
 
     # if getattr(settings, 'MEMCACHE_MECHANISM', False) is True:
